+++ CHANGELOG.md
@@ -2,6 +2,12 @@ TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
-0.1.0 (2021-02-99)
+0.1.1 (2021-03-03)
+=====
+
+* add option: '-p, --pipe-out <num>   read from pipe <num> [unimplemented]'
+* update crate: flood-tide-gen
+
+0.1.0 (2021-02-28)
 =====
 first commit